Move express config to another file
