+++ src/middleware/boilerplates/runner/index.js
@@ -1,22 +1,10 @@
 'use strict';
 
 const { ServiceBroker } = require('moleculer');
-const ApiGatewayService = require('moleculer-web');
 const os = require('os');
-const express = require('express');
-const passport = require('passport');
-const session = require('express-session');
-const cors = require('cors');
-
-const LdpService = require('@semapps/ldp');
-const { Routes: ActivityPubRoutes } = require('@semapps/activitypub');
-
-const getLoginMiddlewares = require('./cas/getLoginMiddlewares');
-const decodeToken = require('./oidc/decodeToken');
-const verifyToken = require('./oidc/verifyToken');
-const configurePassport = require('./cas/configurePassport');
 
 const createServices = require('./createServices');
+const configureExpress = require('./configureExpress');
 const CONFIG = require('./config');
 
 // Broker init
@@ -28,58 +16,7 @@ const broker = new ServiceBroker({
 });
 
 createServices(broker);
-
-const app = express();
-app.use(
-  session({
-    secret: 's€m@pps',
-    maxAge: null
-  })
-);
-
-app.use(cors());
-app.use(passport.initialize());
-app.use(passport.session());
-
-configurePassport(passport).then(() => {
-  app.use('/auth', ...getLoginMiddlewares());
-});
-
-const apiGateway = broker.createService({
-  mixins: [ApiGatewayService],
-  settings: {
-    middleware: true,
-    cors: {
-      origin: '*',
-      exposedHeaders: '*'
-    },
-    routes: [ActivityPubRoutes, LdpService.routes],
-    defaultLdpAccept: 'text/turtle'
-  },
-  methods: {
-    // https://moleculer.services/docs/0.13/moleculer-web.html#Authentication
-    async authenticate(ctx, route, req, res) {
-      try {
-        const token = req.headers.authorization && req.headers.authorization.split(' ')[1];
-        if (token) {
-          await verifyToken(token);
-          const payload = decodeToken(token);
-          console.log('Token payload', payload);
-          ctx.meta.tokenPayload = payload;
-          return Promise.resolve(payload);
-        } else {
-          return Promise.resolve(null);
-        }
-      } catch (err) {
-        console.log('Invalid token');
-        return Promise.reject();
-      }
-    }
-  }
-});
-
-// Use ApiGateway as middleware
-app.use(apiGateway.express());
+const app = configureExpress(broker);
 
 // Start
 broker
@@ -94,7 +31,7 @@ broker
       if (err) {
         console.error(err);
       } else {
-        console.log('express started');
+        console.log('Listening on port 3000');
       }
     });
